requirements.txt
#
# This file is autogenerated by pip3-compile
# To update, run:
#
#    pip3-compile
#
rospkg == 1.3.0
pyyaml >= 5.4
openpyxl == 3.0.9
spatialmath-python == 0.11
swift-sim == 0.10.0
spatialgeometry == 0.2.0
pandas
numpy >= 1.21
matplotlib == 2.1.1
argparse == 1.4.0
PySide2 == 5.15.2
PySide2extn == 1.0.0
roboticstoolbox-python == 0.11.0
sympy == 1.9
numpy-stl == 2.16.3
alphashape
plotly
torch
torchvision
torchaudio